Adding a helper function to free up the clustering data structures
